initial commit
The purpose of this analysis is to simulate the Convolutional neural network and discuss is working in classifying image data. The data used is the fruit 360 dataset, which contains images of fruits and vegetables with the shape of 100x 100. These are in a 3-dimensional format that is RGB(red, green, blue). Also, in this file, different image preprocessing algorithms are also discussed which can prove to increase the accuracy of the deep learning model. The result discusses the performance of the convolutional neural network for classifying the images in the dataset. It also shows the importance of a convolutional neural network for the task specified.